Implement Playlist get, update, delete methods. Minor bug fixes in UserServiceTests.

+++ RidePalApplication/src/main/java/com/example/ridepalapplication/services/UserServiceImpl.java
@@ -1,8 +1,8 @@
 package com.example.ridepalapplication.services;
 
-import com.example.ridepalapplication.exceptions.AuthorizationException;
 import com.example.ridepalapplication.exceptions.EntityDuplicateException;
 import com.example.ridepalapplication.exceptions.EntityNotFoundException;
+import com.example.ridepalapplication.helpers.CheckPermissions;
 import com.example.ridepalapplication.models.User;
 import com.example.ridepalapplication.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
-import static com.example.ridepalapplication.helpers.CheckPermissions.checkIfSameUser;
+import static com.example.ridepalapplication.helpers.CheckPermissions.checkAuthorization;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -65,7 +65,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User updateUser(User loggedUser, User userToBeUpdated) {
-        checkIfSameUser(loggedUser, userToBeUpdated, "update");
+        checkAuthorization(loggedUser, userToBeUpdated, "update other users");
 
         boolean emailExists = true;
 
@@ -73,7 +73,10 @@ public class UserServiceImpl implements UserService {
             emailExists = false;
         }
         if (emailExists) {
-            throw new EntityDuplicateException("User", "email", userToBeUpdated.getEmail());
+            User userWithSameEmail = userRepository.findByEmail(userToBeUpdated.getEmail());
+            if (!userWithSameEmail.equals(userToBeUpdated)) {
+                throw new EntityDuplicateException("User", "email", userToBeUpdated.getEmail());
+            }
         }
 
         return userRepository.save(userToBeUpdated);
@@ -81,7 +84,14 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(User loggedUser, Long id) {
-        checkIfSameUser(loggedUser, id, "delete");
-        userRepository.deleteById(id);
+        CheckPermissions.checkAuthorization(loggedUser, id, "delete other users");
+
+        Optional<User> userToUpdate = userRepository.findById(id);
+        if (userToUpdate.isPresent()) {
+            userRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("User", id);
+        }
     }
+
 }
